Changed edges to use BitSets
This commit changes the representation of edges in the graph from a 2D array of boolean values to a 1D array of BitSets. The representation is still an adjacency matrix, but the BitSets are used to reduce total space. Also, BitSets might be useful for future functionality since they support set operations.