Added code to merge BitSets representing cliques
The Graph class can now merge together two BitSets that represent cliques. Using BitSets will presumably be faster than other ways of merging two cliques since set operations are fast.  Thus, if we have clique A and clique B, we only have to check that the nodes in A - B all have edges to the nodes in B - A.  Likewise, if A - B = A or B - A = B or A is empty or B is empty, we will not have found a (new) clique.