Added code to randomly merge existing cliques
Unsurprisingly, this approach is slow, since there are many cliques with cardinality 3. Merging some of these generates cliques with cardinality 6, but even 10,000,000 iterations usually does not produce larger cliques, since the cardinality 6 cliques are rare in the sea of cardinality 3 cliques.

For any random approach to be useful, it will need to be cleverer about how it selects cliques, preferring at least one of the cliques to be larger.

This commit also fixes a bug in the mergeCliques() method which did not merge disjoint cliques.