New results
By trying to add all (or most) nodes to the best cliques found so far, we can improve the quality of the cliques found.

For brock200_1.clq, this process yields cliques of size 19 out of a possible 21 in less than a second.

Further tuning for iterations and best clique window size could improve size or speed.