Performance tuning
The core of the program is the same, but it's been tuned in a lot of ways (leaving a lot of ugly comments lying around). Also, I no longer do clique merging throughout the process since it was giving mediocre results after the initial merges. Now, once cliques have been merged, the program tries to add random nodes to existing cliques.

The program does run on all graphs in the data set but not always with very impressive results. It also requires the JVM to be invoked with 4 gigs of heap space (-Xmx4g) to get through some of the largest graphs.